re-added image copying
