+++ Config.py
@@ -1,0 +1,3 @@
+# Arguments
+image_dir = '/home/santiago/Proyectos/Cabinet/TestFiles'
+destination_dir = '/home/santiago/Proyectos/Cabinet/SortedFiles'
